+++ doc/mri.rst
@@ -34,6 +34,4 @@ Step by step instructions for using :func:`gui.coregistration`:
    transforms.apply_volume_registration
    transforms.compute_volume_registration
    vertex_to_mni
-   warp_montage
-   warp_montage_volume
    coreg.Coregistration
