+++ mne/__init__.py
@@ -74,8 +74,7 @@ from .source_estimate import (read_source_estimate,
                               extract_label_time_course, stc_near_sensors)
 from .surface import (read_surface, write_surface, decimate_surface, read_tri,
                       get_head_surf, get_meg_helmet_surf, dig_mri_distances,
-                      warp_montage, warp_montage_volume,
-                      get_montage_volume_labels)
+                      warp_montage_volume, get_montage_volume_labels)
 from .morph_map import read_morph_map
 from .morph import (SourceMorph, read_source_morph, grade_to_vertices,
                     compute_source_morph)
